% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Data.R
\name{download_D1Data}
\alias{download_D1Data}
\title{Downloads data from DataOne with metadata}
\usage{
download_D1Data(data_obj, path)
}
\arguments{
\item{data_obj}{(character) An identifier or url for a DataONE object to download.
If the object is a metadata or resource map object, all data associated with that object will downloaded.}

\item{path}{(character) Path to a directory to download data to}
}
\value{
(D1Data) A list of data and metadata
}
\description{
Downloads data from DataOne with metadata
}
\examples{
\dontrun{
obj <- "https://dx.doi.org/10.6073/pasta/edc06bbae6db80e801b6e52253f2ea09"
D1Data <- get_D1Data(obj)

obj <- "https://cn.dataone.org/cn/v2/resolve/urn\%3Auuid\%3A085ffc05-6e5d-4d5d-9b8c-6763e4bf85ac" #DATA
D1Data <- get_D1Data(obj)
}
}
