% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1Data.R
\name{write_D1Data}
\alias{write_D1Data}
\title{Writes D1Data objects to Disk}
\usage{
write_D1Data(D1Data, path = getwd())
}
\arguments{
\item{D1Data}{(D1Data) A result of \code{get_D1Data}.}

\item{path}{(optional/character) Path to directory to write data to.}
}
\description{
Writes D1Data objects to Disk
}
\examples{
\dontrun{
obj <- "https://dx.doi.org/10.6073/pasta/edc06bbae6db80e801b6e52253f2ea09"
D1Data <- get_D1Data(obj)
write_D1Data(D1Data)

obj <- "https://cn.dataone.org/cn/v2/resolve/urn\%3Auuid\%3A085ffc05-6e5d-4d5d-9b8c-6763e4bf85ac" #DATA
D1Data <- get_D1Data(obj)
write_D1Data(D1Data)
}
}
